Add xl!unsetprefix command
Users weren't sure how to unset a channel prefix as it wasn't documented anywhere. Rather than pitch the "none" feature I found in the code, I added a dedicated command and help for this as I think that will be easier on users to figure out.

For some reason, the linq code doesn't seem to like working when something's been removed, but there's still something left. I don't know why.