Silence more unneded chat types and fix list
Discord was interpretting the first line of the aggregated chat types as a syntax highlighting type, which would skip the first entry. Sometimes. Added a newline before it so avoid this as a good practice.

I've also adjusted some more chattypes to skip checking for player link as they can't/won't have one afaik. Keeps the log tidy.